requirements.txt
tensorflow==2.15.0
matplotlib==3.7.2
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
seaborn==0.12.2
pillow==10.0.0
jupyter==1.0.0
ipykernel==6.25.0
keras==2.15.0
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
python-multipart==0.0.6
requests==2.31.0